multinet/requirements.txt
python-arango
graphql-core
